fsck.overlay: check index dir in workdir and fix index feature
The workdir may contian index dir when it was mounted in overlayfs which
index mount options was enabled.

The old kernel do not set index feature on the upper layer when we mount
overlayfs with "index=on", so this patch check and ask user to fix this
feature when we detect index dir in workdir.

This patch do not add index feature to supported feature set because
fsck is not ready to auto check and fix index related objects, so fsck
will still fail if it detect index feature in feature set xattr or find
index dir in workdir.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>