mkfs.overlay: implement mkfs.overlay utility
mkfs.overlay utility is used to create an overlay filesystem in the
underlying directories on the basic file systme.

mkfs.overlay will init an feature set to each layer's root directory.
Note we may get a not fully initialized filesystem when when some lower
layers are read-only or not support xattr.

This path implement a basic frame of mkfs.overlay and does not come into
use now.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>