mkfs.overlay: refuse to make if index/work dir exists
If "work" or "index" directory exist in workdir, it means these user
specified layers may be mounted by overlayfs and contain an existing
overlay "image". We cannot make a new one here because it may lead to
inconsistency of some overlay features in the underlying layers, so
we refuse to make fs and advise user tp run fsck.overlay to make a
consistent overlay "image".

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>