overlay: assign metacopy feature
Metacopy feature is backword incompatible because it may split the
metadata and data of a file when doing copying up. The old kernel
which does not support this feature may corrupt the metacopy file
when modifying it.

This patch assign the metacopy feature like the Linux kernel does,
doesn't change any other logic.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>