fsck.overlay: fix redirect dir feature
Redirect dir feature need to be set on the underlying layer which have
reditect dirs. This patch ask user to fix redirect dir feature if fsck
detect redirect xattr on some of the layer's sub-directories, but it's
just an suggestion now for backword compatibility, so fsck will not
fail if we don't fix it.

This patch also add redirect_dir feature to the supported feature set
bit map because fsck can fix it now.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>