fsck.overlay: fix index feature
Index feature need to be set on the upper root dir if index dir exists
in the workdir. This patch ask user to fix index feature when index dir
exists, but it's just an suggestion now for backword compatibility, so
fsck will not fail if we don't fix it.

Unfortunately, fsck does not support index feature now, so it doesn't
add this feature to the supported feature set bit map, and fsck will
also quit if the index dir is detected.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>